using AbstractSigner instead of Signer
