+++ packages/protocol-kit/src/adapters/ethers/utils/index.ts
@@ -1,4 +1,4 @@
-import { ContractTransactionResponse, Signer } from 'ethers'
+import { ContractTransactionResponse, Signer, Provider } from 'ethers'
 import { EthersTransactionOptions, EthersTransactionResult } from '../types'
 
 export function sameString(str1: string, str2: string): boolean {
@@ -18,4 +18,19 @@ export function toTxResult(
 
 export function isTypedDataSigner(signer: any): signer is Signer {
   return (signer as unknown as Signer).signTypedData !== undefined
+}
+
+/**
+ * Check if the signerOrProvider is compatible with `Signer`
+ * @param signerOrProvider - Signer or provider
+ * @returns true if the parameter is compatible with `Signer`
+ */
+export function isSignerCompatible(signerOrProvider: Signer | Provider): boolean {
+  const candidate = signerOrProvider as Signer
+
+  const isSigntransactionCompatible = typeof candidate.signTransaction === 'function'
+  const isSignMessageCompatible = typeof candidate.signMessage === 'function'
+  const isGetAddressCompatible = typeof candidate.getAddress === 'function'
+
+  return isSigntransactionCompatible && isSignMessageCompatible && isGetAddressCompatible
 }
